relocate mongo test connection and minor edit
