+++ backend/models/organizations/orgs-model.test.js
@@ -1,6 +1,6 @@
 const OrgsModel = require("./orgs-model")
 const { TagsModel } = require("../tags/tags-model")
-const db = require("../db-test-setup")
+const db = require("../../utils/test/db-test-setup")
 
 // Define test instances here
 const testTags = [
