+++ backend/models/projects/projects-model.test.js
@@ -1,6 +1,6 @@
 // const mongoose = require("mongoose")
 const ProjectsModel = require("./projects-model")
-const db = require("../db-test-setup")
+const db = require("../../utils/test/db-test-setup")
 
 // Define test instances here
 const testProject1 = {
