+++ backend/models/readings/readings-model.test.js
@@ -1,6 +1,6 @@
 // const mongoose = require("mongoose")
 // const readingsModel = require("./readings-model")
-const db = require("../db-test-setup")
+const db = require("../../utils/test/db-test-setup")
 
 // Define test instances here
